move internal process tests to separate files
